+++ tests/formatter/input-output-pairs/if-elif-else-comments.in.gd
@@ -1,0 +1,15 @@
+class X:
+	func foo():
+		for i in [1]:
+			if i:
+			# aaa
+				pass
+			# bbb
+			elif i:
+			# ccc
+				pass
+			# ddd
+			else:
+			# eee
+				pass
+			# fff