+++ tests/formatter/input-output-pairs/complex-extends-statements.out.gd
@@ -18,3 +18,12 @@ class E:
 class F:
 	extends E
 	tool
+
+
+class X:
+	class Y:
+		tool
+
+
+class Z:
+	extends X.Y