+++ tests/valid-gd-scripts/dotted_types.gd
@@ -1,0 +1,5 @@
+class X:
+	class Y:
+		tool
+
+var x: X.Y