+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  - Fixed parent block comments in if-elif-else
  - Fixed parentheses handling in formatter
  - Improved fromatter's dedent lookup
+ - Fixed 'extends' bug in grammar
 
 ## [3.2.3] 2020-02-04