+++ CHANGELOG.md
@@ -7,6 +7,7 @@
  - Fixed parentheses handling in formatter
  - Improved fromatter's dedent lookup
  - Fixed 'extends' bug in grammar
+ - Fixed grammar - types allow dots now
 
 ## [3.2.3] 2020-02-04